Extract Base64Str type into separate module & add tests
